Refactored shapes
