+++ analysis_options.yaml
@@ -17,6 +17,7 @@ analyzer:
     override_on_non_overriding_member: error
     must_call_super: error
     unawaited_futures: error
+    non_constant_identifier_names: ignore
 
 linter:
   # The lint rules applied to this project can be customized in the
@@ -45,7 +46,7 @@ linter:
   - avoid_double_and_int_checks
   - avoid_dynamic_calls
   - avoid_empty_else
-  - avoid_equals_and_hash_code_on_mutable_classes
+  # - avoid_equals_and_hash_code_on_mutable_classes
   - avoid_escaping_inner_quotes
   # - avoid_final_parameters
   - avoid_function_literals_in_foreach_calls
