+++ lib/playx_3d_scene.dart
@@ -26,4 +26,4 @@ export 'src/models/scene/skybox/hdr_skybox.dart';
 export 'src/models/scene/skybox/ktx_skybox.dart';
 export 'src/models/scene/skybox/skybox.dart';
 export 'src/models/scene/scene.dart';
-export 'src/models/shapes/shapes.dart';
+export 'src/shapes/shapes.dart';
